refactor(channels/discord): split index/inbound/outbound into focused modules
- routing.ts: resolveAgentId, resolveSessionKey (from inbound.ts)
- stream-buffer.ts: SegmentedStreamBuffer (from outbound.ts)
- message-chunk.ts: chunkDiscordMessage (from outbound.ts)
- react.ts: reactToMessage (from outbound.ts)
- config.ts: resolveGroupPolicy, isDmAllowed, resolveToken, mapGuildsForReceive (from index.ts)
- access-policy.ts: passesAllowlist (from index.ts)
- stop-command.ts: STOP_CMD_RE, maybeHandleStop (from index.ts)

No behavior change. Test files split alongside.
